refactor(maintenance): move retention kernels to owners
