+++ crates/tracedecay-code-index-retention/Cargo.toml
@@ -11,6 +11,7 @@ repository = "https://github.com/ScriptedAlchemy/tracedecay"
 cap-fs-ext = "4.0.2"
 cap-std = "4.0.2"
 fs2 = "0.4"
+gix = { version = "=0.86.0", default-features = false, features = ["revision", "parallel", "sha1", "sha256", "status"] }
 hotpath.workspace = true
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
@@ -26,3 +27,4 @@ libc = "0.2"
 
 [dev-dependencies]
 tempfile = "3"
+tracedecay-runtime-core = { path = "../tracedecay-runtime-core", version = "0.1.0" }
